fix: ratchet dojo debt to 623
