+++ docs/design/WF-155-respecting-dojo-ratchet-624.md
@@ -38,6 +38,15 @@ Live counts on `main` at `5fa0aea9`:
 - code-size baseline: `55`, including `4` legacy hard-limit files
 - next aggregate target: `624` or lower
 
+Implemented counts on this branch before review:
+
+- aggregate Code Dojo debt: `623`
+- ESLint findings: `215`
+- file/context baseline: `331`
+- mock-ban baseline: `22`
+- code-size baseline: `55`, including `4` legacy hard-limit files
+- next aggregate target: `573` or lower
+
 Top current ESLint offender clusters:
 
 - `examples/docs/storybook-app.ts`: `12`
@@ -48,9 +57,30 @@ Top current ESLint offender clusters:
 - `examples/docs/storybook-workstation.ts`: `5`
 - `packages/bijou-tui/src/runtime-engine.ts`: `5`
 
+Selected cleanup slice:
+
+- `examples/docs/storybook-app.ts`
+- `examples/docs/storybook-workstation.ts`
+- `examples/docs/counter-block-demo.ts`
+- `packages/bijou-i18n/src/localization.test.ts`
+- `packages/bijou-tui-app/src/index.ts`
+- `packages/bijou-tui-app/src/index.test.ts`
+- `packages/bijou-tui/src/collection-surface.ts`
+- `packages/bijou-tui/src/focus-area.test.ts`
+- `packages/bijou/src/core/components/surface-primitives.test.ts`
+- `examples/split-editors/main.ts`
+- `packages/bijou-node/src/io.ts`
+
 The DOGFOOD app entrypoint remains oversized after the previous split:
 
 - `examples/docs/app.ts`: `4,165` physical lines / `147,799` bytes
+
+This branch keeps cutting that file while preserving visible docs behavior:
+
+- `examples/docs/app.ts`: `3,306` physical lines / `114,890` bytes
+- extracted standard block docs, block preview surfaces, and Theme Inspector
+  state into focused sub-150-line modules
+- DOGFOOD raw-string debt: `2,373`, down from the `2,415` baseline
 
 ## Scope
 
@@ -138,7 +168,7 @@ without re-running broad discovery first.
 ## Acceptance Criteria
 
 - The selected touched files reduce live findings by at least `50`.
-- Aggregate Code Dojo debt is `624` or lower.
+- Aggregate Code Dojo debt is `623` or lower.
 - The ESLint baseline records the lower live count.
 - The Code Dojo exception ledger and `package.json` report the lower ceiling.
-- The next ratchet target is documented as `574` or lower.
+- The next ratchet target is documented as `573` or lower.
