+++ examples/docs/app-standard-block-docs.ts
@@ -1,0 +1,4 @@
+export { standardBlockDocumentationText } from './app-standard-block-documentation-docs.js';
+export { standardBlockInventoryMarkdown } from './app-standard-block-inventory-docs.js';
+export { standardBlockLoweringMarkdown } from './app-standard-block-lowering-docs.js';
+export { standardBlockPreviewMarkdown } from './app-standard-block-preview-docs.js';
