+++ examples/docs/i18n/source/dogfood-strings.csv
@@ -1,4 +1,8 @@
 namespace,id,kind,sourceLocale,sourceValueKind,sourceValue,locale,valueKind,value,status,description
+bijou.dogfood,blocks.preview.appShell.navigation,message,en,string,Guides / Components / Blocks,de,string,Guides / Komponenten / Blocks,current,
+bijou.dogfood,blocks.preview.appShell.navigation,message,en,string,Guides / Components / Blocks,en,string,Guides / Components / Blocks,current,
+bijou.dogfood,blocks.preview.appShell.navigation,message,en,string,Guides / Components / Blocks,es,string,Guias / Componentes / Bloques,current,
+bijou.dogfood,blocks.preview.appShell.navigation,message,en,string,Guides / Components / Blocks,fr,string,Guides / Composants / Blocs,current,
 bijou.dogfood,blocks.preview.activityStream.event.prOpened,message,en,string,10:42 PR opened,de,,,missing,
 bijou.dogfood,blocks.preview.activityStream.event.prOpened,message,en,string,10:42 PR opened,en,string,10:42 PR opened,current,
 bijou.dogfood,blocks.preview.activityStream.event.prOpened,message,en,string,10:42 PR opened,es,,,missing,
@@ -211,6 +215,18 @@ bijou.dogfood,blocks.preview.inFlowStatus.source,message,en,string,DOGFOOD Block
 bijou.dogfood,blocks.preview.inFlowStatus.source,message,en,string,DOGFOOD Blocks,en,string,DOGFOOD Blocks,current,
 bijou.dogfood,blocks.preview.inFlowStatus.source,message,en,string,DOGFOOD Blocks,es,,,missing,
 bijou.dogfood,blocks.preview.inFlowStatus.source,message,en,string,DOGFOOD Blocks,fr,,,missing,
+bijou.dogfood,blocks.preview.inspectorPanel.action.focus,message,en,string,Focus source,de,string,Quelle fokussieren,current,
+bijou.dogfood,blocks.preview.inspectorPanel.action.focus,message,en,string,Focus source,en,string,Focus source,current,
+bijou.dogfood,blocks.preview.inspectorPanel.action.focus,message,en,string,Focus source,es,string,Enfocar fuente,current,
+bijou.dogfood,blocks.preview.inspectorPanel.action.focus,message,en,string,Focus source,fr,string,Cibler la source,current,
+bijou.dogfood,blocks.preview.inspectorPanel.action.reveal,message,en,string,Reveal selection,de,string,Auswahl anzeigen,current,
+bijou.dogfood,blocks.preview.inspectorPanel.action.reveal,message,en,string,Reveal selection,en,string,Reveal selection,current,
+bijou.dogfood,blocks.preview.inspectorPanel.action.reveal,message,en,string,Reveal selection,es,string,Mostrar seleccion,current,
+bijou.dogfood,blocks.preview.inspectorPanel.action.reveal,message,en,string,Reveal selection,fr,string,Afficher la selection,current,
+bijou.dogfood,blocks.preview.inspectorPanel.selection,message,en,string,ReaderSurface,de,string,ReaderSurface,current,
+bijou.dogfood,blocks.preview.inspectorPanel.selection,message,en,string,ReaderSurface,en,string,ReaderSurface,current,
+bijou.dogfood,blocks.preview.inspectorPanel.selection,message,en,string,ReaderSurface,es,string,ReaderSurface,current,
+bijou.dogfood,blocks.preview.inspectorPanel.selection,message,en,string,ReaderSurface,fr,string,ReaderSurface,current,
 bijou.dogfood,blocks.preview.inlineStatus.label,message,en,string,Docs inventory,de,,,missing,
 bijou.dogfood,blocks.preview.inlineStatus.label,message,en,string,Docs inventory,en,string,Docs inventory,current,
 bijou.dogfood,blocks.preview.inlineStatus.label,message,en,string,Docs inventory,es,,,missing,
@@ -343,6 +359,22 @@ bijou.dogfood,blocks.preview.progressIndicator.label,message,en,string,Install p
 bijou.dogfood,blocks.preview.progressIndicator.label,message,en,string,Install packages,en,string,Install packages,current,
 bijou.dogfood,blocks.preview.progressIndicator.label,message,en,string,Install packages,es,,,missing,
 bijou.dogfood,blocks.preview.progressIndicator.label,message,en,string,Install packages,fr,,,missing,
+bijou.dogfood,blocks.preview.readerSurface.content,message,en,string,ReaderSurface live content from DOGFOOD Blocks.,de,string,ReaderSurface Live-Inhalt aus DOGFOOD Blocks.,current,
+bijou.dogfood,blocks.preview.readerSurface.content,message,en,string,ReaderSurface live content from DOGFOOD Blocks.,en,string,ReaderSurface live content from DOGFOOD Blocks.,current,
+bijou.dogfood,blocks.preview.readerSurface.content,message,en,string,ReaderSurface live content from DOGFOOD Blocks.,es,string,Contenido vivo de ReaderSurface desde DOGFOOD Blocks.,current,
+bijou.dogfood,blocks.preview.readerSurface.content,message,en,string,ReaderSurface live content from DOGFOOD Blocks.,fr,string,Contenu live ReaderSurface depuis DOGFOOD Blocks.,current,
+bijou.dogfood,blocks.preview.readerSurface.navigation,message,en,string,Blocks navigation,de,string,Blocks Navigation,current,
+bijou.dogfood,blocks.preview.readerSurface.navigation,message,en,string,Blocks navigation,en,string,Blocks navigation,current,
+bijou.dogfood,blocks.preview.readerSurface.navigation,message,en,string,Blocks navigation,es,string,Navegacion de Blocks,current,
+bijou.dogfood,blocks.preview.readerSurface.navigation,message,en,string,Blocks navigation,fr,string,Navigation Blocks,current,
+bijou.dogfood,blocks.preview.readerSurface.outline.lowering,message,en,string,How Blocks Lower,de,string,Wie Blocks senken,current,
+bijou.dogfood,blocks.preview.readerSurface.outline.lowering,message,en,string,How Blocks Lower,en,string,How Blocks Lower,current,
+bijou.dogfood,blocks.preview.readerSurface.outline.lowering,message,en,string,How Blocks Lower,es,string,Como bajan los Blocks,current,
+bijou.dogfood,blocks.preview.readerSurface.outline.lowering,message,en,string,How Blocks Lower,fr,string,Comment les Blocks descendent,current,
+bijou.dogfood,blocks.preview.readerSurface.outline.what,message,en,string,What are Blocks,de,string,Was sind Blocks,current,
+bijou.dogfood,blocks.preview.readerSurface.outline.what,message,en,string,What are Blocks,en,string,What are Blocks,current,
+bijou.dogfood,blocks.preview.readerSurface.outline.what,message,en,string,What are Blocks,es,string,Que son los Blocks,current,
+bijou.dogfood,blocks.preview.readerSurface.outline.what,message,en,string,What are Blocks,fr,string,Que sont les Blocks,current,
 bijou.dogfood,blocks.preview.progressiveDisclosure.detail.debugTraces,message,en,string,debug traces,de,string,Debug-Spuren,current,
 bijou.dogfood,blocks.preview.progressiveDisclosure.detail.debugTraces,message,en,string,debug traces,en,string,debug traces,current,
 bijou.dogfood,blocks.preview.progressiveDisclosure.detail.debugTraces,message,en,string,debug traces,es,string,trazas de depuración,current,
